Create a world module for world model and sector
